:new: wCommunity algorithm
